:wrench: Finalise script
